Removed OnyxWarmthController, renamed OnyxSDMController to OnyxWarmthController
